Move more JSPs, actions, and classes
