+++ .gitattributes
@@ -3414,17 +3414,7 @@ study/src/org/labkey/study/view/requirePipeline.jsp -text
 study/src/org/labkey/study/view/saveReportView.jsp -text
 study/src/org/labkey/study/view/sendParticipantGroup.jsp -text
 study/src/org/labkey/study/view/snapshotSettings.jsp -text
-study/src/org/labkey/study/view/specimen/labSpecimenLists.jsp -text
-study/src/org/labkey/study/view/specimen/manageActorOrder.jsp -text
-study/src/org/labkey/study/view/specimen/manageActors.jsp -text
-study/src/org/labkey/study/view/specimen/manageRequest.jsp -text
-study/src/org/labkey/study/view/specimen/manageRequestInputs.jsp -text
-study/src/org/labkey/study/view/specimen/manageRequestStatus.jsp -text
-study/src/org/labkey/study/view/specimen/manageRequirement.jsp -text
-study/src/org/labkey/study/view/specimen/manageStatuses.jsp -text
-study/src/org/labkey/study/view/specimen/manageStatusOrder.jsp -text
 study/src/org/labkey/study/view/specimen/updateComments.jsp -text
-study/src/org/labkey/study/view/specimen/webPart.jsp -text
 study/src/org/labkey/study/view/studydesign/assayScheduleWebpart.jsp -text
 study/src/org/labkey/study/view/studydesign/AssayScheduleWebpartFactory.java -text
 study/src/org/labkey/study/view/studydesign/immunizationScheduleWebpart.jsp -text
@@ -3457,10 +3447,8 @@ study/src/org/labkey/study/visitmanager/SequenceVisitManager.java -text
 study/src/org/labkey/study/visitmanager/VisitManager.java -text
 study/src/org/labkey/study/visualization/DataspaceVisualizationProvider.java -text
 study/src/org/labkey/study/visualization/StudyVisualizationProvider.java -text
-study/src/org/labkey/study/writer/AssayDatasetWriter.java -text
 study/src/org/labkey/study/writer/AssayScheduleWriter.java -text
 study/src/org/labkey/study/writer/CohortWriter.java -text
-study/src/org/labkey/study/writer/DatasetWriter.java -text
 study/src/org/labkey/study/writer/DefaultStudyDesignWriter.java -text
 study/src/org/labkey/study/writer/InternalStudyWriter.java -text
 study/src/org/labkey/study/writer/MissingValueWriterFactory.java -text
