+++ specimen/src/org/labkey/specimen/actions/IdForm.java
@@ -1,4 +1,4 @@
-package org.labkey.api.specimen.actions;
+package org.labkey.specimen.actions;
 
 import org.labkey.api.action.ReturnUrlForm;
