+++ specimen/src/org/labkey/specimen/actions/ManageRequestInputsBean.java
@@ -1,4 +1,4 @@
-package org.labkey.api.specimen.actions;
+package org.labkey.specimen.actions;
 
 import org.labkey.api.data.Container;
 import org.labkey.api.specimen.SpecimenRequestManager;
