+++ specimen/src/org/labkey/specimen/actions/SpecimenEventBean.java
@@ -1,4 +1,4 @@
-package org.labkey.api.specimen.actions;
+package org.labkey.specimen.actions;
 
 import org.labkey.api.action.ReturnUrlForm;
 import org.labkey.api.specimen.Vial;
