+++ specimen/src/org/labkey/specimen/view/manageRequestInputs.jsp
@@ -16,10 +16,10 @@
  */
 %>
 <%@ page import="org.labkey.api.specimen.SpecimenRequestManager.SpecimenRequestInput"%>
-<%@ page import="org.labkey.api.specimen.actions.ManageRequestInputsBean"%>
 <%@ page import="org.labkey.api.study.StudyUrls"%>
 <%@ page import="org.labkey.api.view.HttpView"%>
 <%@ page import="org.labkey.api.view.JspView"%>
+<%@ page import="org.labkey.specimen.actions.ManageRequestInputsBean"%>
 <%@ page import="org.labkey.specimen.actions.SpecimenController2.HandleUpdateRequestInputsAction" %>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
